Bug fix, millisecond timestamp format
It was caused by the float to string casting of millisecond values together with pd.to_datetime() microsecond format (%f).
When the millisecond value contained only 2 digits before the decimal, it was not zero-padding when converting to microseconds in string format. 
This led to timestamps with two-digit millisecond values to be 10x the millisecond count while the three-digit values are correct.